Add view.commands.rm

+++ src/controller/command/rm.py
@@ -2,6 +2,7 @@
 # =====================================================
 # Internal
 import model.db as db
+import view.commands.rm as view
 
 
 # SECTION HANDLER - Delete a job from the app.
@@ -9,6 +10,8 @@ import model.db as db
 def handler(command, argument):
   current_job = db.get_state()['job']
   if argument == current_job:
-    # return view.cannotDeleteCurrentJob(current_job)
-    return False
-  return db.delete(argument)
+    return view.cannotDeleteCurrentJob(current_job)
+  if db.exists(argument):
+    db.delete(argument)
+    return view.deleted(argument)
+  view.notFound(argument)
